requirements-dev.txt
pytest>=3.4.0
ipython>=6.2.1
ipdb>=0.10.3
flask_debugtoolbar>=0.10.1
pytest_flask>=0.10.0
codecov>=2.0.15
pytest-cov>=2.5.1
flake8>=3.5.0
flake8-bugbear>=18.2.0
isort>=4.3.3
twine>=1.9.1
wheel>=0.30.0
flask>=1.0
logzero
pytest_mock
mock
pytest_django
pyyaml